feat(fix): document type filtering
